Added device commands to the arduino over Serial
You are now able to send serial commands to the Circuit Playground Express that will control some behavior. In this commit, it’s the ability to turn on some of the NeoPixels and make the fade in & out, either for a set duration, or indefinitely. Commands also allow for the effect to be terminated. See the circuitPlayground/README.md for more details.

Signed-off-by: Cameron Perry <cameron.perry@valtech.com>